Clean up and writing tests
