+++ app/src/utils/apiClients.tsx
@@ -242,7 +242,6 @@ export class CollectionsApi {
     }/search/?q=${encodeURIComponent(
       keyword
     )}${filterURL}&sort=${sort}&page=${page}&perPage=${perPage}`;
-    console.log(apiUrl);
 
     const response = await fetchApi({
       apiUrl: apiUrl,
